Improve the LiveTests a bit
For the live tests, we now simply assert that the failure rate is below
some threshold. If the failure rate is too high, we also print some
statistics about how often we've observed a particular error. Finally,
if VERBOSE is enabled, we print every failing sample to stdout.

Drive-By: make the MockFuzzer use the same CodeGenerator weights as the
real fuzzer.
Drive-By: disallow splicing during value building as it has a higher
probability of leading to invalid code.