Track minimization overhead
With this change, we now compute a "minimization overhead" for our
statistics which represents the fraction of (recent) executions spent on
minimization.
For that, we add a |purpose| field to every execution which describes
why we're executing a program. For the statistics we now only use the
.minimization purpose, but if we want to, we could extend this in the
future to get a detailed breakdown of exactly what our executions are
spent on.