Change the RegExp logic
Split the well-known patterns into patterns that work with the u and v
flag and into patterns that don't work with these flags.
As wrong regular expressions lead to syntax errors it is not possible to
guard these with try-catch statements. Therefore we try to reduce the
number of wrong regular expressions we emit.