Add infrastructure for post-processing samples during fuzzing
We now allow registering a "post-processor" with a FuzzingEngine. If
registerd, the post-processor will be invoked for every generated sample
and can modify it arbitrarily (by returning a different program). This
can for example be useful in situations where certain actions should be
performed at (more or less) random places in the generated programs.