Make single explore operations deterministic.
Add a JS RNG such that we can seed explore operations.
If Fuzzilli is unable to convert explore operations into concrete
operations when trying to reproduce a crash, the explore code itself
might be important. In such cases we still want this explore operation
to behave the same way it did when we observed the crash.