Use setrlimit instead of relying on the core_pattern
Use setrlimit(2) to disable core dumps in the child process, instead of
forcing a core_pattern across the whole system. Specifically, set
RLIMIT_CORE to 0.

Drive by change: Increase allowed failure rate of the LiveTests,
this is higher than it is with V8 as the NodeJS version of Github is
different and does not support the same features (i.e. Regular
Expression flags).