Switch to a JavaScriptExecutor class
Switch from a specific NodeJS class used for tests and
the FuzzIL compiler to a more generic JavaScriptExecutor
that uses the binary given through the FUZZILLI_TEST_SHELL
environment variable to execute tests.
The FuzzIL compiler also uses this class although it still needs
npm modules and therefore still tries to find the NodeJS binary
in the users path.