Lift a WasmModule to itself instead of its exports
We now lift a WasmModule to itself, and then access its exports
property with a regular `GetProperty` Instruction.
This way the Explore Mutator can access the full module and additionally
our Sandbox fuzzer can find the full module instead of its exports.
We now explicitly type the output of `GetProperty` during generation.
We will lose type information as soon as the initial sample is mutated.