Remove ProgramBuilder::Mode
Historically, the ProgramBuilder supported two modes: .aggressive
(mostly used for the MutationEngine) and .conservative (mostly used for
the HybridEngine). The main distinction used to be the behavior of
the various variable selection routines: in conservative mode, they
would guarantee to find a variable of the right type (or fail), in
aggressive mode they could return a variable of a different type.
However, with the introduction of the new variable selection APIs
(randomVariable(ofType:) and randomVariable(forUseAs:) in particular),
the builder mode no longer has any influence on the behavior of variable
selection. Instead, it is currently only used for two, somewhat niche
use cases: during splicing to decide between randomVariable(ofType:) and
randomVariable(forUseAs:) when searching for existing variables, and in
one code generator, which is skipped in conservative mode. These two
probably do not warrant keeping support for different modes in the
ProgramBuilder, so with this change, the mode is now gone.