Change Minimization and Indices of Instructions.
Change the way we track indices of instructions: instead of having an
extra inout, we now track the index as an attribute on the Instruction.
We also add a Flags struct to the Instruction class which allows us to
mark Instructions as not-removable by the minimizer.
This is necessary as some Reducers in the Minimizer move code around,
which invalidates any indices that are tracked, the new flags help us
keep track of instructions we initially marked as not-removable.
All changes to code by Reducers now must go through the Minimization
helper which provides some Methods to do so.