Makes break context identification lightweight
These changes move the logic for break context identification out of the performance ciritical part of the Fuzzilli project and into the compiler.

The idea is straight-forward: We fix the bug in the context analyzer that caused contexts to contain both .loop and .switchBlock by simply removing the loop context upon opening a switchBlock context and vice versa.