Introduce WasmTypeExtension, adjust globals to WasmGlobalTypeExtension
The ILType.object can get a WasmTypeExtension structure that is meant to
encode Wasm-specific type information. This is used to pass info from JS
to the Typer.

This change also introduced the WasmGlobalTypeExtension that encodes
the underlying type and the mutability of a Wasm global.
The subsumption rules will be adjusted in a follow-up CL.