Add some debugging helpers
Programs that fail during LiveTests are now dumped to disk if one passes
FUZZILLI_TEST_DEBUG=1 to swift test.
Programs now have a method to store them to a disk, which can be helpful
during debugging, and there is also a new convienience initializer that
loads Programs from a given path.