Remove --random-seed argument from V8Profile
In a typical deployment, multiple Fuzzilli instances will cooperate and
send crashing samples to a dedicated "master" instance. In such a
configuration, it doesn't make much sense for each instance to pick a
random --random-seed as all instances of such a "swarm" need to use the
same seed. As such, with this change we remove the argument from the
V8Profile. Instead, the user can pass a --random-seed argument via the
--additionalArguments and that way ensure that the same seed is used.