Fixes Switch Break Bug
Bug can be seen when trying to compile Tests/FuzzilliTests/CompilerTests/switch_statements.js, which fails because "break statement outside of loop or switch", even though it isn't. This happens because breaks are expected in the switch cases, rather than the switch block. This was missed because compilation was succesful before making the Analyzer changes a seperate branch and verification of the new ContextAnalyzer was only done with Tests/FuzzilliTests/AnalyzerTest.swift and not with Tests/CompilerTests/switch_statements.js