Implements temporary fix
The problem that is addressed with this commit can be seen when using testFunctionLifting. The test fails because the expected f0(a1) is instead f0() - i.e. somehow the parameter a1 wasn't lifted. This happens because the patterns array in the new Parameter array was initialized empty and only populated afterwards. In a later commit we will make patterns a constructor parameter and immutable afterwards.