Simplify WasmLifter
We now pass a typer into the WasmLifter that re-types the JavaScript
program during lifting. This is now a typer that might have less
information than we had during program building as any setType operation
in a template will be lost. This does enable us to query the typer for
the types of variables that we need to import. Specifically we can now
easily use the WasmTypeExtensions to get additional information that we
need during lifting. We can therefore also get rid of the import
generators.