[environment] Allow method overloads in ObjectGroup
So far only a single signature was allowed per method name, but in
reality, with method overloads there can be several possible signatures.
This CL changes the method field of an ObjectGroup to not store a map
from method name to signature, but a map from method name to an array of
signatures.