Make it possible to import .js files that throw an exception
Previously, we wouldn't import those files as we require all samples in
the corpus to execute successfully. With this change, we simply try to
wrap all failing programs in a big try-catch block during corpus import,
then try importing them again. In the future we could improve this
mechanism to for example determine exactly which instruction causes the
exception and then either remove this instruction or wrap (only) it in a
try-catch. But for now, this simple change already allows us to import
~30% more programs from V8's regression tests.

Drive-By: add some more functions to filteredFunctionsForCompiler.