Improve corpus import fixup further
With the preceeding change, we can now import failing programs by
wrapping them in a big try-catch statement. This is somewhat
inefficient, however, as we may end up with samples where a lot of the
code is never executed (because an exception is thrown early on). This
change now adds additional logic to improve that and ensure that as much
code as possible will also be executed at runtime. Specifically, the
fixup algorithm when importing programs now works like this:

1. We first try to replace known test functions (such as `assertEquals`)
   with a dummy function as these test functions aren't available
   outside the testing environment. This now replaces a mechanism in the
   compiler that would also try to remove calls to these functions.

2. If the first attempt fails, we'll then attempt to insert try-catch
   blocks around individual instructions. For that, we first enable all
   guardable operations (which will cause them to be wrapped in
   try-catch), then perform one round of fixup during which all
   unecessary guards will be disabled again. As a result, only those
   try-catch block that are necessary will stay in the imported program.

3. Only if the previous attempts fail do we now insert one big try-catch
   statement around the entire program.