Simplify lifting of wasm signatures
This is a preparation to allow blocks with non-void signatures.
This means that there is yet another construct registering signatures.
As the calculation of signature indices has become quite convoluted,
this change adds a mechanism to register a siganture and then later on
the signature can be used to query its corresponding index.

This change also performs deduplication of equivalent signatures as a
side effect.