Support subtyping in FindOrGenerateType
The JavaScriptEnvironment has a dictionary that stores for every type a
set of methods that produce that type. With this CL, the dictionary now
also contains for a type the set of methods that produce a subtype of
that type.

There is a second dictionary that contains properties that return a
type. The same changes were done for that dictionary.

Change-Id: I8971ebb41a95906be4b5c8c5b997abe6f06fd3b4
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/7910526
Commit-Queue: Andreas Haas <ahaas@google.com>
Reviewed-by: Samuel Groß <saelo@google.com>