Improve handling of block bodies in parser
We now have a common helper function (visitBody) to process the body of
various statements instead of repeating the same code fragment.
Similarly, we now also have a common visitParameters function. Finally,
this change also adds a test to make sure that we don't introduce
additional block statements due to the way statement bodies are
represented in the AST. See the comment in the test for more details.

Change-Id: Id494669c02ec36132daf7327bf2654fc21e13e0e
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/7934187
Reviewed-by: Carl Smith <cffsmith@google.com>
Commit-Queue: Samuel Groß <saelo@google.com>