Give functions names
This change introduces a new intermediate FuzzIL operation class:
BeginAnyNamedFunction. This class has an optional functionName string
and is used for plain, generator, async, and async generator functions.
During lifting, if the functionName is present, it will be used as the
function's identifier instead of an automatically assigned name
(something like `f23`). This now makes it possible to correctly compile
function definitions. In particular, it makes it possible to support
test cases that rely on function hoisting as the function's name is
important for that to work.

In general, we still prefer function with "dynamic" names (where the
lifter assigns a unique identifier during lifting) to avoid name
collisions during mutations (e.g. because we're splicing a function with
the same name as an existing function into a program). As such, this
change also modifies the minimizer to try and remove function names if
they are not important for the behavior of the program.

Change-Id: Ia6169d4440c7de3a77c38f71edfc2a728456076d
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/7934267
Commit-Queue: Samuel Groß <saelo@google.com>
Reviewed-by: Carl Smith <cffsmith@google.com>