Introduce DataFlowSimplifier
This is a generic reducer that attempts to simplify the data flow of a
program. For example, for the following program:

  v3 <- Foo
  v4 <- Bar v3
  v5 <- Baz v4

And assuming that only `Baz` is the relevant operation, it would attempt
to simplify it to just:

  v3 <- Foo
  v5 <- Baz v3

This should generally be a useful simplification, but is in particular
helpful to remove named variables from the program if the names are not
important. In the future, we could also likely replace the
special-purpose ReassignReducer with this generic reducer.

Change-Id: I402451e3ad4b301381ba4929d28a008f7ac93670
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/7934188
Reviewed-by: Carl Smith <cffsmith@google.com>
Commit-Queue: Samuel Groß <saelo@google.com>