[wasm] Add support for blocks with result
This changes wasm blocks (only the "plain" block, not ifs, trys etc.)
to have an optional result type. A wasm block with a result produces
a Variable in its WasmEndBlock operation. The value is either the input
argument to the WasmEndBlock (in case of regularly reaching the end of
the block) or the arguments passed into a WasmBranch or WasmBranchIf
operation that branches to the WasmBeginBlock's label.

As these branch instructions need to provide the argument(s) matching
the block signature, the label type is extended to also contain type
information about the arguments to be passed along a branch.

Change-Id: I4a1af782330c7a71320a2d4b1949dba081d1d4de
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/7967831
Reviewed-by: Carl Smith <cffsmith@google.com>
Commit-Queue: Matthias Liedtke <mliedtke@google.com>