[wasm] Restructure try-catch blocks in Fuzzilli
Previously, the wasm try catch would look somewhat like this in FuzzIL:
  try -> L:v1
    throw v2
    v3 = 123
    catch v2 -> v4
      // v3 is visible here but it may not be initialized.
    endcatch
    // the code generator might still insert more regular wasm
    // statements here...
  endtry

This CL changes that to:
  try -> L:v1
    throw v2
    v3 = 123
  catch v2 -> L:v4 v5
    // v3 is not visible here.
  endtry

where the catch block finishes the try block and starts a new block at
the same time (for the catch content). The endtry therefore is an end
block for either `try`, `catch`, or `catch_all`.

Change-Id: If61b1c4fd99c630c1fb879cd4db310d7c5509018
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/7967834
Reviewed-by: Carl Smith <cffsmith@google.com>
Commit-Queue: Matthias Liedtke <mliedtke@google.com>