Make WasmLifting robust
We now wrap WasmLifting in a try block such that we can catch some
errors where type information is incorrect. This can happen for a
multitude of reasons, one being the MutationEngine changing some
operation to produce a different type. As following code has made some
assumptions about the input that no longer hold, we need to abort.
We now also throw an exception in JavaScript, such that we don't keep
that sample around (except when the Wasm module itself is in a
try-catch).

Change-Id: Id0a781a04cc78505d299646147eb0eb0bb9a3aa4
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/7959748
Commit-Queue: Carl Smith <cffsmith@google.com>
Reviewed-by: Matthias Liedtke <mliedtke@google.com>