[improve livetests] Limit max generated value of randomNonNegativeIndex()
This helper function is used to generate random indices for memory
accesses. In some cases, it calls the randomSize() function which has an
upper cap of the generated number which is way larger than the Wasm
memory size usually. This was not an issue, as we used the remainder
after modulo memory_size.
However, as the generated number is uniformly chosen from a set of
interesting integers which are close to the multiples of the memory_size,
too often we ended up using values close to the memory_size.
To fix this issue, we will pass the memory_size to randomSize() as the
upper cap, so no multiples are generated, therefore, we will not get the
memory_size value often.

Change-Id: I157508ee5f9f1a6453de99f8066fd655d67547cb
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/7963788
Reviewed-by: Carl Smith <cffsmith@google.com>
Commit-Queue: Eva Herencsárová <evih@google.com>