[types] Add enum type
Enums (called enumerations, to avoid a conflict with the swift keyword
"enum") are implemented as strings with a TypeExtension. The list of all
possible values of the enum are encoded in the "properties" field of the
TypeExtension.

Additionally, ProgramBuilder.findOrGenerateType is extended to only pick
the allowed values in case the requested type is an enumeration.

Change-Id: Ic9021314b9b884daffb6f47468032051755369fd
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/7970028
Reviewed-by: Carl Smith <cffsmith@google.com>
Commit-Queue: Andreas Haas <ahaas@google.com>