[wasm] Introduce WasmOperationBase and legacy WasmTypedOperation
For wasm-gc and for multi-output instructions, the current
WasmOperation is too inflexible.
Instead with this CL a WasmOperationBase is introduced as a generic
base class for all wasm operations. Differently to the WasmOperation
(which is renamed to WasmTypedOperation), the typing of the outputs is
done by custom code in the JSTyper for each wasm operation.

This is closer to how typing works in JS and allows defining multiple
output values as well as deriving the output ILType based on the inputs
(which is needed for wasm-gc operations as the resulting ILType of an
ArrayNewFixed depends on the array type that was passed into this
operation as an input.

Change-Id: Id6ad02924e74471148d647172f2d2ffd1401971b
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8061128
Reviewed-by: Carl Smith <cffsmith@google.com>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>