[wasm] Make ILType.wasmFuncRef and .wasmExternRef wasm-gc references
This introduces the concept of abstract wasm heap types in a way that
will be compatible in the future with other wasm-gc reference types.

This change also adds the so far unused exnref abstract heap type
which is part of the (new) exception-handling proposal.

Change-Id: I3fcc6e2f5b84c946c2446bfc4c0221ba3356c64d
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8074489
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Carl Smith <cffsmith@google.com>