[wasm] Add globals with exnref and externref element type
Note that this only adds exnref as an element type to globals defined
in wasm as it is not allowed to define globals of exnref in JS via
new WebAssembly.Global().

Globals of externref can be defined both in JS as well as in wasm.

Change-Id: I147b042ac7c30810a5b2758c958cbc641e54ce38
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8107670
Reviewed-by: Carl Smith <cffsmith@google.com>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>