Enable CodeGenMutator to mutate TypeGroups
With this change the CodeGenMutator is now aware of TypeGroups.
We don't CodeGen on regular type definition instructions anymore, as
those would be discarded as unused type defintions anyways since they
are not wired up to the EndTypeGroup instruction.

On EndTypeGroup instruction we now generate code before adopting that
instruction with updated inputs and outputs.

drive by fix for the OperationMutator handling ExnRef and ExternRef
WasmGlobal types.

Change-Id: I98dfe0e1c04eae5725879924ac2713b74663edec
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8116948
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Carl Smith <cffsmith@google.com>