[wasm] Fix memory leak for self and cyclic forward references
For wasm-gc types due to the option of self-referencing its own type or
forward-referencing another type (that could then reference the type
with the forward-reference to form a cycle), the ILTypes form similar
cycles. This results in memory leaks as swift doesn't have a GC.

Solution: To break these cycles, the WasmReferenceType extension only
stores a weak (unowned) reference to the description of an indexed
type. The lifetime is controlled by the WasmTypeDefinition extension
which is always present (otherwise we couldn't construct a reference
type to it).

Note that this limits the usability of such ILType to the lifetime of
the JSTyper which keeps all the other ILTypes (and therefore also the
WasmTypeDefinition extensions) alive.

Change-Id: I7ad5a01bdb55ea3688ae904eb697506f7cb96302
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8135230
Reviewed-by: Carl Smith <cffsmith@google.com>
Commit-Queue: Matthias Liedtke <mliedtke@google.com>