[wasm] Remove isMemory64 property from memory operations
The memory knows its type and we can ask the JSTyper for it, no need
for the operations using the memory to track this information
separately. (This also aligns the operations better with WasmMemorySize
and WasmMemoryGrow that also don't have a property for isMemory64.

Change-Id: I807ef736f7078eb248e53547eb8d86b20df17dd2
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8155437
Auto-Submit: Matthias Liedtke <mliedtke@google.com>
Reviewed-by: Carl Smith <cffsmith@google.com>
Commit-Queue: Carl Smith <cffsmith@google.com>