Fix a rare failure to generate a type for a known Signature.
The testArgumentGenerationForKnownSignature test would rarely fail
because it wants to generate an ArrayBuffer, for the ArrayBuffer our
generation logic could figure out that it can create an e.g. Int8Array
and access its buffer property. Since the Int8Array is also an
.iterable, our CodeGneeration logic would then randomly also chose to
generate an iterable by creating a regular Array and then it would try
to access the 'buffer' which then did not exist and failed to fulfil the
test assertion.

Change-Id: I5838ced36dd9b2030253bdcdb0ebf26d8233c77d
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8184716
Reviewed-by: Matthias Liedtke <mliedtke@google.com>
Commit-Queue: Carl Smith <cffsmith@google.com>