Improve Statistics for CodeGenerators
Alongside their Correctness rate (how often the generated code has led
to a correct sample if it has added instructions) we now also collect
the rate of how often it has generated code when it was invoked.
A CodeGenerator might get called but fail to emit instructions because
of some dynamic requirements that are not met. In this case we print how
often it emitted something vs how often it was invoked.

Change-Id: I18f176db191f4901032b5d9a343e3744e2673609
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8184717
Commit-Queue: Carl Smith <cffsmith@google.com>
Reviewed-by: Samuel Groß <saelo@google.com>