Make Wasm importAnalysis bail on mistyped input
Mutations can reassign to objects in JS such as
.object(ofGroup: "WasmTable"), if we already have code that relies on it
being a WasmTable and it is reassigned to be a .function() our
importAnalysis will try to analyze it as a .function and see it in an
instruction it does not expect a .function in. In that case we also need
to bail, as that instruction does not have the expected input.

Change-Id: I48a41f8900812a2e5eaaef51482d7f91bf150382
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8307419
Reviewed-by: Matthias Liedtke <mliedtke@google.com>
Commit-Queue: Matthias Liedtke <mliedtke@google.com>
Auto-Submit: Carl Smith <cffsmith@google.com>