Implement compilation of class expressions
This enables JS class expressions to be parsed and compiled. The
implementation reuses parts of the existing machinery for class
declarations as the body of such definitions is the same. Internally,
class declarations are already treated as if they were expressions
(i.e. the class value is stored in a variable). The main difference
is that the name in class expressions is optional.

This change does not yet differentiate between declarations and
expressions internally, compilations is a mere program transformation
that turns class expressions into FuzzIL class definitions. On
lifting, those become declarations. In a follow up, we could implement
storing this information and lifting it differently.

Implementation and review guidance:
Patch https://crrev.com/i/8381596/6 is a simple refactoring, where
two methods are extracted without altering their code. This prepares
sharing their code in the next patch.
Patch https://crrev.com/i/8381596/6..7 implements parsing and
compilation.

When importing mjsunit regression tests, this increases compiling
samples from 2474 to 2495 of 2795.

It fixes the 37 errors of type "Unhandled node ClassExpression",
out of which 21 are sufficiently fixed and 16 shift to other
error types (e.g. other not yet supported class constructs within
the class bodies).

Bug: 428620828
Change-Id: I0a730351d6fcd8494ad1fd84611c1b0d99694918
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8381596
Reviewed-by: Matthias Liedtke <mliedtke@google.com>
Reviewed-by: Samuel Groß <saelo@google.com>
Commit-Queue: Michael Achenbach <machenbach@google.com>