Propagate class-expression information
Previously we added the isExpression flag to class definitions in
the FuzzIL compiler. This change propagates the information to
the operation and lifting. JS lifting now creates class expressions
and assigns them to variables.

Additionally this adds class expressions in program generators
that previously created declarations with a probability of 30%.

Bug: 428620828
Change-Id: I9a6c41f89a4c427b1dd34edece07e4179164f547
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8407077
Reviewed-by: Matthias Liedtke <mliedtke@google.com>
Reviewed-by: Carl Smith <cffsmith@google.com>
Commit-Queue: Michael Achenbach <machenbach@google.com>