Always emit variables for unused expressions if configured
To generate better test cases for differential fuzzers, with this
change Fuzzilli can be configured to always emit variables for
expressions, even if they are unused.
Note that this requires a differential fuzzer to still instrument the
test cases to print or otherwise expose the defined variables.

Bug: 422361840
Change-Id: I26b72261887e0c493653e1020b59d567835f8dcc
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8433337
Reviewed-by: Michael Achenbach <machenbach@google.com>
Commit-Queue: Matthias Liedtke <mliedtke@google.com>