[wasm] Fix crash of tag without .wasmTagType
This fixes a crash in Fuzzilli where we have the following pattern:
  v0 <- CreateWasmTag [.wasmi32]
  v1 <- CreateWasmTag [.wasmi64]
  v2 <- LoadBoolean 'true'
  v3 <- TernaryOperation v2, v0, v1

The type of v3 is the union(type(v0), type(v1)) which resulted in
having a .object(ofGroup: "WasmTag") as both inputs are tags but not
having a WasmTypeExtension as these are incompatible (two different)
tag types.

When the WasmTypeExtension gets invalidated by the union, we now also
remove the object group from the TypeExtension.

Change-Id: I00da60a29974e768d7321c577b0015cdd31f8029
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8444056
Reviewed-by: Samuel Groß <saelo@google.com>
Reviewed-by: Carl Smith <cffsmith@google.com>
Commit-Queue: Matthias Liedtke <mliedtke@google.com>