Add V8SandboxFuzzer
This is a basic fuzzer for the V8 Sandbox. It uses the memory corruption
API to implement a random-but-deterministic (given a seed) traversal
through the V8 heap object graph and corrupts some objects found this
way in a deterministic way. It will also log all important decisions to
produce a fairly readable log of the steps leading to a crash.

Change-Id: Ic7ff03d3cb8b85230f07a5db2f8442c01f32fda4
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8463484
Commit-Queue: Samuel Groß <saelo@google.com>
Reviewed-by: Matthias Liedtke <mliedtke@google.com>