[wasm] Add WebAssembly.<function> builtins to JavaScriptEnvironment
Wasm fuzzing in Fuzzilli is done by high-level operations in the FuzzIL
that get lowered to various calls on the WebAssembly object in
JavaScript, like new WebAssembly.Module().
As these high-level FuzzIL operations will only ever generate valid
WebAssembly modules and are not meant to fuzz the WebAssembly builtins
themselves but rather the Wasm <-> JS interop, adding the builtins
themselves to the environment should help Fuzzilli to also fuzz these
JS interfaces in a less-constrained mode (meaning it will probably not
be able to ever generate a WebAssembly.validate() call that returns
true, which however also isn't the goal here.)

This CL adds the functions to validate, compile and instantiate a Wasm
module. As a next step, all the constructors need to be added.

Bug: 440226707
Change-Id: Icdb574c307278536360da347bd200a288725819d
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8526836
Commit-Queue: Matthias Liedtke <mliedtke@google.com>
Reviewed-by: Manos Koukoutos <manoskouk@google.com>