[temporal] Handle relativeTo options
Temporal has a relativeTo option that shows up in a couple places.

Unlike our solution for union signatures
(https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8549330/10)
this shows up inside an OptionsBag. We could still do that type of
solution, but then we'd need a large number of registered OptionsBag
types for each subtype, which is somewhat messy.

I'm not a huge fan of making jsTemporalRelativeTo special in OptionsBag,
but it works. Another option would be to teach the producing generators
code to detect types like that. It is not a simple `.object()` type
anymore, so the existing code in `generateTypeInternal()` shouldn't work
(I think?). But we could potentially make it work.


Bug: 439921647
Change-Id: I6a6a69640b8a2110f240a71b59fd2a4824ba2be8
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8560946
Reviewed-by: Matthias Liedtke <mliedtke@google.com>
Commit-Queue: Manish Goregaokar <manishearth@google.com>