[temporal] Bias towards using a generator for temporal types
Otherwise we end up generating massive chains of Temporal methods, e.g.
"make a date and a datetime, call until on the date with the datetime,
then call add on a third datetime with the result of that, then call
toPlainYearMonth, then ..."

We want to test these things, but not that much: ApiMethodCallGenerator
should suffice for most of the chained method testing.

This should eliminate the "Reached variable generation limit in
generateType" warnings we keep seeing.

Bug: 439921647
Change-Id: I6a6a6964015ce0b114f039295a5ef7bc471761a1
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8579369
Auto-Submit: Manish Goregaokar <manishearth@google.com>
Reviewed-by: Matthias Liedtke <mliedtke@google.com>
Commit-Queue: Matthias Liedtke <mliedtke@google.com>