Implement a new CodeGeneration approach.
CodeGenerators are now annotated with the types they produce and the
Contexts they provide. This allows us to pick an (almost) arbitrary
CodeGenerator and nest it in others such that we can definitely execute
it. Add a ContextGraph that makes this assembly possible. Because we now
have explicit yield points in CodeGenerators we can remove the
buildRecursive method.

Bug: 408966832
Change-Id: I0320bdd5fa64bd687cea87075b6607ceb61d779d
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8386801
Reviewed-by: Matthias Liedtke <mliedtke@google.com>
Commit-Queue: Carl Smith <cffsmith@google.com>