[Turbolev] Adjust fuzzing flags probabilities
After a discussion, we are adjusting the flag probabilities.

The stable version of Turbolev (--turbolev) is now enabled with a 50% probability. In 82% of those cases, the future version (--turbolev-future) is also enabled, as a result it is never used alone.

Change-Id: I8b19b0f16414fd7928040020c9b4045820ef6b41
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8609477
Commit-Queue: Dominik Klemba <tacet@google.com>
Auto-Submit: Dominik Klemba <tacet@google.com>
Reviewed-by: Matthias Liedtke <mliedtke@google.com>
Reviewed-by: Darius Mercadier <dmercadier@google.com>