Support computed class methods
This adds support for computed class method (static and instance)
to fuzzIL and to the compiler for imports.

Supporting computed getters/setters and indexed methods + indexed
getters/setters is still future work. However, this adds a compiler
test file with the full array of cases that we could support,
not-yet-supported cases commented out for now.

This slightly refactors the compiler AST and shares property keys
for methods and properties, which can have the same structure.

In general, most of the boiler-plate code is similar to the existing
code for BeginClass{Instance|Static}Method with changes for computed
properties similar to BeginClassComputedProperty or
BeginObjectLiteralComputedMethod.

Similar to computed object methods, a caveat is that the typer might
not be able to compute available method names to be called.

Bug: 446634535
BYPASS_LARGE_CHANGE_WARNING: tests for completeness are large
Change-Id: I30a85eadbc0fe466a9fbc7dbfa3bb7aad5b51f7d
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8614819
Reviewed-by: Matthias Liedtke <mliedtke@google.com>
Commit-Queue: Michael Achenbach <machenbach@google.com>