New codegen: Don't reroll to splice on each 'level'
Previously we had ProgramBuilder.buildRecursive() to recursively
generate code into a program which was replaced with
ProgramBuilder.build() calls with the new code generation mechanism.

However the default BuildingMode is .generatingAndSplicing which now
rerolls a die on each build(n: x) call in a code generator whether to
generate or splice which significantly increases the amount of splicing
done by Fuzzilli.

As splicing is very expensive and involves rewriting the whole program,
we should not randomly splice when running CodeGenerators. For
ProgramTemplates (similarly to before) each build call still rolls a die
whether to generate or splice (however, when deciding to generate on a
build() call, this decision will then "stick" as code generators will be
using .generating in all cases now.

Bug: 448862171
Change-Id: I982c0f992994d2592de43b71c2b59f3dd334e2bb
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8637338
Commit-Queue: Matthias Liedtke <mliedtke@google.com>
Reviewed-by: Carl Smith <cffsmith@google.com>