Improve class-definition generation
Previously there were two code generators for classes, the
ClassDefinitionGenerator and the WeirdClassGenerator.

The former was still an old-style generator, so it could not be
scheduled for any of the ~20 generators that require the
classDefinition context.

So all such generators scheduled the WeirdClassGenerator, which
forced a simple method as superclass, which in return doesn't contain
any interesting code as in e.g. the ObjectConstructorGenerator.

This change unifies the two class generators as one new-style
generator. The generator provides the classDefinition context, so it
can be scheduled for parts of classes. It also produces a
constructor() (unlike the WeirdClassGenerator), so it can be
scheduled for generators that require constructor()-type input.

Bug: 446632644
Change-Id: I7313dae3998cdd0ff488779debeca7c6ab333fc9
Reviewed-on: https://chrome-internal-review.googlesource.com/c/v8/fuzzilli/+/8705459
Commit-Queue: Michael Achenbach <machenbach@google.com>
Reviewed-by: Matthias Liedtke <mliedtke@google.com>